+++ lib/Provider/SearchProvider.php
@@ -77,12 +77,11 @@ class SearchProvider extends \OCP\Search\Provider
         
         if ($query !== null) {
             
-            $solrResult = $this->solrService->search($query);
+            $solrResult = $this->solrService->search($query, array(
+                ''
+            ));
             if ($solrResult == false)
                 return $results;
-            
-            if (sizeof($solrResult) > 0)
-                $topScore = $solrResult[0]['score'];
             
             foreach ($solrResult as $data) {
                 
@@ -94,7 +93,14 @@ class SearchProvider extends \OCP\Search\Provider
                 $result = new \OC\Search\Result\File($fileData);
                 $result->type = 'nextant';
                 // $result->name = $result->path . ' (Accuracy: ' . round($data['score'] * 100 / $topScore, 2) . '%) ';
-                $result->name = $result->path . ' (Score: ' . round($data['score'], 2) . ') ';
+                
+                $name = '';
+                $name .= ($data['owner'] != $this->userId) ? '[shared] ' : '';
+                $name .= ($data['deleted']) ? '[trashbin] ' : '';
+                $name .= $result->path;
+                $name .= ' (Score: ' . round($data['score'], 2) . ') ';
+                
+                $result->name = $name;
                 $results[] = $result;
             }
         }