+++ lib/Service/ConfigService.php
@@ -35,7 +35,8 @@ class ConfigService
         'configured' => '1',
         'needed_index' => '1',
         'solr_url' => 'http://127.0.0.1:8983/solr/',
-        'solr_core' => 'nextant'
+        'solr_core' => 'nextant',
+        'solr_lock' => 0
     ];
 
     private $appName;