+++ lib/Service/SolrAdminService.php
@@ -25,6 +25,7 @@
  */
 namespace OCA\Nextant\Service;
 
+use \OCA\Nextant\Service\SolrService;
 use Solarium\Core\Client\Request;
 
 class SolrAdminService
@@ -36,8 +37,6 @@ class SolrAdminService
 
     private $miscService;
 
-    private $output;
-
     private $lastMessage;
 
     public function __construct($solrService, $configService, $miscService)
@@ -46,12 +45,6 @@ class SolrAdminService
         $this->solrService = $solrService;
         $this->configService = $configService;
         $this->miscService = $miscService;
-        $this->output = null;
-    }
-
-    public function setOutput(&$output)
-    {
-        $this->output = $output;
     }
 
     public function checkSchema($fix = false, &$error = '')
@@ -63,48 +56,80 @@ class SolrAdminService
         
         $fields = array();
         array_push($fields, array(
-            'name' => 'nextant_owner',
-            'type' => 'string',
-            'indexed' => true,
-            'stored' => true,
-            'multiValued' => false
+            'type' => 'dynamic-field',
+            'data' => array(
+                'name' => 'nextant_attr_*',
+                'type' => 'text_general',
+                'indexed' => true,
+                'stored' => true,
+                'multiValued' => true
+            )
         ));
         array_push($fields, array(
-            'name' => 'nextant_share',
-            'type' => 'string',
-            'indexed' => true,
-            'stored' => true,
-            'multiValued' => true
+            'type' => 'field',
+            'data' => array(
+                'name' => 'nextant_owner',
+                'type' => 'string',
+                'indexed' => true,
+                'stored' => true,
+                'multiValued' => false
+            )
         ));
         array_push($fields, array(
-            'name' => 'nextant_sharegroup',
-            'type' => 'string',
-            'indexed' => true,
-            'stored' => true,
-            'multiValued' => true
+            'type' => 'field',
+            'data' => array(
+                'name' => 'nextant_mtime',
+                'type' => 'int',
+                'indexed' => true,
+                'stored' => true,
+                'multiValued' => false
+            )
         ));
         array_push($fields, array(
-            'name' => 'nextant_deleted',
-            'type' => 'boolean',
-            'indexed' => true,
-            'stored' => false,
-            'multiValued' => false
+            'type' => 'field',
+            'data' => array(
+                'name' => 'nextant_share',
+                'type' => 'string',
+                'indexed' => true,
+                'stored' => true,
+                'multiValued' => true
+            )
+        ));
+        array_push($fields, array(
+            'type' => 'field',
+            'data' => array(
+                'name' => 'nextant_sharegroup',
+                'type' => 'string',
+                'indexed' => true,
+                'stored' => true,
+                'multiValued' => true
+            )
+        ));
+        array_push($fields, array(
+            'type' => 'field',
+            'data' => array(
+                'name' => 'nextant_deleted',
+                'type' => 'boolean',
+                'indexed' => true,
+                'stored' => true,
+                'multiValued' => false
+            )
         ));
         
-        $this->message('Checking Solr schema fields');
+        $this->solrService->message('Checking Solr schema fields');
         
         $changed = false;
         while (true) {
             foreach ($fields as $field) {
-                $this->message(' * field ' . $field['name'] . ': ', false);
+                $this->solrService->message(' * Checking ' . $field['type'] . ' \'' . $field['data']['name'] . '\': ', false);
                 if (self::checkFieldProperty($client, $field, $curr))
-                    $this->message('ok.');
+                    $this->solrService->message('ok.');
                 else {
-                    $this->message('fail. ');
+                    $this->solrService->message('fail. ');
                     
                     if ($fix) {
                         $changed = true;
-                        $this->message('   -> Fixing field ' . $field['name']);
+                        $this->solrService->message('   -> Fixing ' . $field['type'] . ' \'' . $field['data']['name'] . '\'');
                         if ($curr)
                             self::modifyField($client, $field);
                         else
@@ -119,7 +144,7 @@ class SolrAdminService
         
         if ($changed)
             $this->configService->setAppValue('needed_index', '1');
-            
+        
         $this->configService->setAppValue('configured', '1');
         return true;
     }
@@ -137,11 +162,11 @@ class SolrAdminService
             return true;
         } catch (\Solarium\Exception\HttpException $ehe) {
             if ($ehe->getStatusMessage() == 'OK')
-                $error = SolrClient::EXCEPTION_SOLRURI;
+                $error = SolrService::EXCEPTION_SOLRURI;
             else
-                $error = SolrClient::EXCEPTION_HTTPEXCEPTION;
+                $error = SolrService::EXCEPTION_HTTPEXCEPTION;
         } catch (\Solarium\Exception $e) {
-            $error = SolrClient::EXCEPTION;
+            $error = SolrService::EXCEPTION;
         }
         
         return false;
@@ -149,25 +174,33 @@ class SolrAdminService
 
     private static function checkFieldProperty($client, $field, &$property)
     {
-        $property = self::getFieldProperty($client, $field['name']);
+        $property = self::getFieldProperty($client, $field['type'], $field['data']['name']);
         if (! $property)
             return false;
         
-        $k = array_keys($field);
+        $k = array_keys($field['data']);
         foreach ($k as $key) {
-            if ($field[$key] != $property[$key])
+            if ($field['data'][$key] != $property[$key])
                 return false;
         }
         
         return true;
     }
 
-    private static function getFieldProperty($client, $fieldName)
+    private static function getFieldProperty($client, $fieldType, $fieldName)
     {
+        $url = '';
+        if ($fieldType == 'field')
+            $url = 'schema/fields/';
+        if ($fieldType == 'dynamic-field')
+            $url = 'schema/dynamicfields/';
+        if ($url == '')
+            return false;
+        
         $query = $client->createSelect();
         $request = $client->createRequest($query);
         
-        $request->setHandler('schema/fields/' . $fieldName);
+        $request->setHandler($url . $fieldName);
         
         $response = $client->executeRequest($request);
         if ($response->getStatusCode() != 200)
@@ -185,7 +218,7 @@ class SolrAdminService
     private static function createField($client, $field)
     {
         $data = array(
-            'add-field' => $field
+            'add-' . $field['type'] => $field['data']
         );
         return self::solariumPostSchemaRequest($client, $data);
     }
@@ -193,16 +226,16 @@ class SolrAdminService
     private static function modifyField($client, $field)
     {
         $data = array(
-            'replace-field' => $field
+            'replace-' . $field['type'] => $field['data']
         );
         return self::solariumPostSchemaRequest($client, $data);
     }
 
-    private static function deleteField($client, $fieldName)
+    private static function deleteField($client, $field)
     {
         $data = array(
-            'delete-field' => array(
-                'name' => $fieldName
+            'delete-' . $field['type'] => array(
+                'name' => $field['data']['name']
             )
         );
         return self::solariumPostSchemaRequest($client, $data);
@@ -227,11 +260,11 @@ class SolrAdminService
             return true;
         } catch (\Solarium\Exception\HttpException $ehe) {
             if ($ehe->getStatusMessage() == 'OK')
-                $error = SolrClient::EXCEPTION_SOLRURI;
+                $error = SolrService::EXCEPTION_SOLRURI;
             else
-                $error = SolrClient::EXCEPTION_HTTPEXCEPTION;
+                $error = SolrService::EXCEPTION_HTTPEXCEPTION;
         } catch (\Solarium\Exception $e) {
-            $error = SolrClient::EXCEPTION;
+            $error = SolrService::EXCEPTION;
         }
         return false;
     }
@@ -253,11 +286,11 @@ class SolrAdminService
             return true;
         } catch (\Solarium\Exception\HttpException $ehe) {
             if ($ehe->getStatusMessage() == 'OK')
-                $error = SolrClient::EXCEPTION_SOLRURI;
+                $error = SolrService::EXCEPTION_SOLRURI;
             else
-                $error = SolrClient::EXCEPTION_HTTPEXCEPTION;
+                $error = SolrService::EXCEPTION_HTTPEXCEPTION;
         } catch (\Solarium\Exception $e) {
-            $error = SolrClient::EXCEPTION;
+            $error = SolrService::EXCEPTION;
         }
         
         return false;
@@ -279,25 +312,14 @@ class SolrAdminService
             return $resultset->getNumFound();
         } catch (\Solarium\Exception\HttpException $ehe) {
             if ($ehe->getStatusMessage() == 'OK')
-                $error = SolrClient::EXCEPTION_SOLRURI;
+                $error = SolrService::EXCEPTION_SOLRURI;
             else
-                $error = SolrClient::EXCEPTION_HTTPEXCEPTION;
+                $error = SolrService::EXCEPTION_HTTPEXCEPTION;
         } catch (\Solarium\Exception $e) {
-            $error = SolrClient::EXCEPTION;
+            $error = SolrService::EXCEPTION;
         }
         
         return false;
-    }
-
-    private function message($line, $newline = true)
-    {
-        if ($this->output != null) {
-            if ($newline)
-                $this->output->writeln($line);
-            else
-                $this->output->write($line);
-        } else
-            $this->lastMessage = $line;
     }
 }