+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+**v0.3.4**
+* halt long update process
+
+**v0.3.3**
+* Exception issue in SolrAdmin/SolrTools
+* rework on the indexing process
+
+**v0.3.2**
+* better script to index your current documents
+* better managment of all file operations (move, shares, trash, ...)
+
+
 **v0.3.1**
 * 'configured' flag to verify that the app is well configured.
 * auto-check your solr setup on upgrade.